Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/mediacapture-fromelement/capture-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: [object DOMError]
CONSOLE MESSAGE: Unhandled Promise Rejection: [object DOMError]
CONSOLE MESSAGE: Unhandled Promise Rejection: [object DOMError]

PASS video-only
FAIL Untitled video.captureStream is not a function. (In 'video.captureStream()', 'video.captureStream' is undefined)
PASS audio-only
FAIL Untitled 1 video.captureStream is not a function. (In 'video.captureStream()', 'video.captureStream' is undefined)
PASS video+audio
FAIL Untitled 2 video.captureStream is not a function. (In 'video.captureStream()', 'video.captureStream' is undefined)
